rename from ags to amg
