--- src/ags/README.rst
@@ -1,5 +1,0 @@
-Microsoft Azure CLI 'ags' Extension
-==========================================
-
-This package is for the 'ags' extension.
-i.e. 'az ags'
