+++ src/amg/HISTORY.rst
