+++ src/amg/README.rst
@@ -1,0 +1,5 @@
+Microsoft Azure CLI 'amg' Extension
+==========================================
+
+This package is for the 'amg' extension.
+i.e. 'az grafana'
