+++ src/amg/azext_amg/_client_factory.py
@@ -3,8 +3,8 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-def cf_ags(cli_ctx, *_):
+def cf_amg(cli_ctx, *_):
     # pylint: disable=unused-argument
     from azure.cli.core.commands.client_factory import get_mgmt_service_client
-    from azext_ags.vendored_sdks import DashboardManagementClient
+    from azext_amg.vendored_sdks import DashboardManagementClient
     return get_mgmt_service_client(cli_ctx, DashboardManagementClient)
