+++ src/amg/azext_amg/tests/latest/test_amg_scenario.py
@@ -1,0 +1,41 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+import os
+import unittest
+
+from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer)
+
+
+TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
+
+
+class AgsScenarioTest(ScenarioTest):
+
+    @ResourceGroupPreparer(name_prefix='cli_test_amg')
+    def test_ags(self, resource_group):
+
+        self.kwargs.update({
+            'name': 'test1'
+        })
+
+        self.cmd('grafana show -g yugangw -n yugangweus2e')
+
+        #self.cmd('ags create -g {rg} -n {name} --tags foo=doo', checks=[
+        #    self.check('tags.foo', 'doo'),
+        #    self.check('name', '{name}')
+        #])
+        #self.cmd('ags update -g {rg} -n {name} --tags foo=boo', checks=[
+        #    self.check('tags.foo', 'boo')
+        #])
+        #count = len(self.cmd('ags list').get_output_in_json())
+        #self.cmd('ags show - {rg} -n {name}', checks=[
+        #    self.check('name', '{name}'),
+        #    self.check('resourceGroup', '{rg}'),
+        #    self.check('tags.foo', 'boo')
+        #])
+        #self.cmd('ags delete -g {rg} -n {name}')
+        #final_count = len(self.cmd('ags list').get_output_in_json())
+        #self.assertTrue(final_count, count - 1)
