+++ src/amg/azext_amg/vendored_sdks/_vendor.py
