+++ src/amg/azext_amg/vendored_sdks/_version.py
