+++ src/amg/azext_amg/vendored_sdks/aio/_patch.py
