+++ src/amg/azext_amg/vendored_sdks/models/_models_py3.py
