+++ src/amg/get_start.md
@@ -13,37 +13,37 @@ The test script below should explain command flows. Everything is under “az gr
     set -e
     RG=test
     LOCATION=westeurope
-    AGS=ygtestgrafana
+    AMG=ygtestgrafana
     az group create -g $RG -l $LOCATION
 
     # Create a grafana instance. Managed Identity is enabled by default. Roles assignments are created for both the command user and managed identity
     # Use "--skip-system-assigned-identity" and "--skip-role-assignments" to skip them, and use "az role assignment create" for finer control
-    az grafana create -g $RG -n $AGS
-    az grafana show -n $AGS
+    az grafana create -g $RG -n $AMG
+    az grafana show -n $AMG
 
     # wait for 2 minutes for RBAC to propagate the change
     sleep 2m
 
     # Create a folder
     FOLDER=testSQL
-    folder_id=$(az grafana folder create -n $AGS --title $FOLDER --query "id" -otsv)
-    az grafana folder show -n $AGS --folder $folder_id
+    folder_id=$(az grafana folder create -n $AMG --title $FOLDER --query "id" -otsv)
+    az grafana folder show -n $AMG --folder $folder_id
 
     # Configure a Azure SQL data source and test the connection
-    data_source_id=$(az grafana data-source create -n $AGS --definition /mnt/d/work/cli/data-source-sql.json --query "id" -o tsv)
-    az grafana data-source show -n $AGS --data-source $data_source_id
-    az grafana data-source query -n $AGS --data-source $data_source_id --query-format table  --conditions rawSql="SELECT 1"
+    data_source_id=$(az grafana data-source create -n $AMG --definition /mnt/d/work/cli/data-source-sql.json --query "id" -o tsv)
+    az grafana data-source show -n $AMG --data-source $data_source_id
+    az grafana data-source query -n $AMG --data-source $data_source_id --query-format table  --conditions rawSql="SELECT 1"
 
     # Configure data dashboard under the folder
-    data_dashboard_uid=$(az grafana dashboard create -n $AGS --folder $FOLDER --title "test dashboard" --definition /mnt/d/work/cli/dashboard-sql.json --query "uid" -o tsv)
+    data_dashboard_uid=$(az grafana dashboard create -n $AMG --folder $FOLDER --title "test dashboard" --definition /mnt/d/work/cli/dashboard-sql.json --query "uid" -o tsv)
     ## Please note the "=" between arg name and parameter. This is to avoid a CLI bug caused by potential leading "-" in uid
-    az grafana dashboard show -n $AGS --dashboard="$data_dashboard_uid"
+    az grafana dashboard show -n $AMG --dashboard="$data_dashboard_uid"
 
     ## tear down
-    az grafana dashboard delete -n $AGS --dashboard="$data_dashboard_uid"
-    az grafana folder delete -n $AGS --folder $FOLDER
+    az grafana dashboard delete -n $AMG --dashboard="$data_dashboard_uid"
+    az grafana folder delete -n $AMG --folder $FOLDER
 
-    az grafana delete -n $AGS --yes
+    az grafana delete -n $AMG --yes
 ```
 
 Notes:
